norm: Fix norm Fix general norm error, except for assert and comment for future check. - Extract ft_pwd() from ft_env.c file to new ft_pwd.c file.
