+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/15 15:39:14 by sanghupa          #+#    #+#             */
-/*   Updated: 2023/09/23 12:04:01 by sanghupa         ###   ########.fr       */
+/*   Updated: 2023/09/23 16:32:02 by sanghupa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -268,6 +268,9 @@ void	ft_export(t_sent *node, t_elst *lst);
 
 /*src/built-in/ft_unset */
 void	ft_unset(t_sent *node, t_elst *lst);
+
+/*src/built-in/ft_pwd */
+void	ft_pwd(t_sent *node, t_elst *lst);
 
 /// readcmd
 /* src/readcmd/readcmd.c */
