+++ src/built-in/ft_unset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_unset.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: minakim <minakim@student.42berlin.de>      +#+  +:+       +#+        */
+/*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/21 14:40:09 by minakim           #+#    #+#             */
-/*   Updated: 2023/09/21 14:42:44 by minakim          ###   ########.fr       */
+/*   Updated: 2023/09/23 16:34:31 by sanghupa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
