+++ src/util/ft_strcspn.c
@@ -6,7 +6,7 @@
 /*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/18 19:59:27 by minakim           #+#    #+#             */
-/*   Updated: 2023/09/13 00:08:46 by minakim          ###   ########.fr       */
+/*   Updated: 2023/09/23 16:19:09 by sanghupa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,6 @@
 
 /// @note These functions have been 'briefly' tested.
 size_t	ft_strcspn(const char *str, const char *charset);
-
 
 /// @brief Calculates the length of the initial segment of a string that does
 /// NOT match with characters from a specified set.
