+++ src/util/ft_strequ.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strqu.c                                         :+:      :+:    :+:   */
+/*   ft_strequ.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: minakim <minakim@student.1337.ma>          +#+  +:+       +#+        */
+/*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/04 16:10:17 by minakim           #+#    #+#             */
-/*   Updated: 2023/09/02 12:33:19 by minakim          ###   ########.fr       */
+/*   Updated: 2023/09/23 16:19:17 by sanghupa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
